Added Italian translation in translation.json
My attempt at an Italian translation, with a couple of things to clear:
Line 99: I don't know if "Queue" is used as a verb or as a noun here, so I translated it as a noun. The verb version would be "Aggiungi alla coda".
Line 102: If this needs to be short, "Succ" really is the best option {It is short for "Successivo").
Line 103: Again, I don't know if "Filter" is used as a noun or as a verb, so I went with verb. The noun would be "Filtro".
There could be some mistakes due to the fact that Italian has gendered nouns and verbs that differ if they refer to a single item or more than one, like in line 122, but that is something that will come to light only when seen in context.